+++ resources/views/content/stock/stocks/edit.blade.php
@@ -1,5 +1,6 @@
 @extends('layouts/app')
 @section('title', $breadcrumbs[count($breadcrumbs) - 1]['name'])
+@inject('provider', 'App\Http\Controllers\Function\GlobalController')
 
 @section('content')
     @include('layouts/panels/breadcrumb', ['breadcrumbs' => $breadcrumbs])
@@ -14,10 +15,12 @@
 
                     <div class="card-body">
                         <div class="mb-2 d-flex justify-content-start justify-content-md-end d-print-none">
-                            <button type="button" onclick="deleteData({{ $query->id }})" class="btn btn-danger mr-2">
-                                <i class="fa fa-trash"></i>
-                                <span class="ms-2">Delete</span>
-                            </button>
+                            @if ($provider::access('adjustment')->access[0]->can_delete == 1)
+                                <button type="button" onclick="deleteData({{ $query->id }})" class="btn btn-danger mr-2">
+                                    <i class="fa fa-trash"></i>
+                                    <span class="ms-2">Delete</span>
+                                </button>
+                            @endif
 
                             <button type="button" onclick='backToList()' class="btn btn-dark">
                                 <i class="fa fa-share"></i>
@@ -180,10 +183,13 @@
                         <div class="row mt-4">
                             <div class="col-12">
                                 <div class="d-flex justify-content-end d-print-none">
-                                    <button class="btn btn-primary me-2" type="submit">
-                                        <i class="fa fa-save"></i>
-                                        <span class="ms-2">Save</span>
-                                    </button>
+                                    @if ($provider::access('adjustment')->access[0]->can_update == 1)
+                                        <button class="btn btn-primary me-2" type="submit">
+                                            <i class="fa fa-save"></i>
+                                            <span class="ms-2">Save</span>
+                                        </button>
+                                    @endif
+                                    
                                     <button type="button" onclick='cancelTrans({{ $query->id }})' class="btn btn-dark">
                                         <i class="fa fa-refresh"></i>
                                         <span class="ms-2">Cancel</span>
